% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bact_coastal_contact_rec.R
\name{Coastal_Contact_rec}
\alias{Coastal_Contact_rec}
\title{Bacteria Coastal Contact Recreation Analysis}
\usage{
Coastal_Contact_rec(df, datetime_column = "sample_datetime")
}
\arguments{
\item{df}{dataframe with Ecoli data}

\item{datetime_column}{POSIXCT column name containing sample datetimes}
}
\value{
a dataframe with relevant Ecoli criteria and excursion variables added
}
\description{
Assesses Ecoli data against the standard
}
\examples{
function(df = your_ecoli_data, datetime_column = "sample_datetime")
}
