% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_assessment.R
\name{DO_30DADMean_calc}
\alias{DO_30DADMean_calc}
\title{30 Day Average Daily Mean DO saturation percent calculator}
\usage{
DO_30DADMean_calc(station, start_date, end_date, df, n = 29)
}
\arguments{
\item{station}{The station to calculate}

\item{start_date}{The first day in the 30 day period}

\item{df}{The dataframe containing DO sat values for subsetting}

\item{n}{The number of days required in the 30 day period}

\item{end}{The the last day in the 30 day period}
}
\description{
This function will calculate 30DADMean DO saturation percentage 
based on daily mean DO sat values. It takes a station, start date, 
end date (totaling 30 days), and a dataframe with mean DO sat values.
}
\examples{
DO_30DADMean_calc(station, start_date, end_date, df, n = 29)

}
