% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_assessment.R
\name{DO_assessment}
\alias{DO_assessment}
\title{DO Analysis}
\usage{
DO_assessment(df, datetime_column = "sample_datetime",
  spawn_start_column = "spawn_start", spawn_end_column = "spawn_end",
  result_column = "Result_cen", temp_column = "temperature",
  elev_column = "ELEV_Ft")
}
\arguments{
\item{df}{dataframe with DO data, including temperature and elevation columns}

\item{datetime_column}{POSIXCT column name containing sample datetimes}

\item{spawn_start_column}{mm/dd spawn start date (as characters)}

\item{spawn_end_column}{mm/dd spawn end date (as characters)}

\item{result_column}{numeric results column name}

\item{temp_column}{column with numeric temperature values in degrees celsius (NAs for no data)}

\item{elev_column}{column with numeric elevation values}
}
\value{
a dataframe with relevant Ecoli criteria and excursion variables added
}
\description{
Assesses dissolved oxygen data against the relevant standard and calculates DO sat where applicable
}
