% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursions_required.R
\name{excursions_req}
\alias{excursions_req}
\title{excursions_req}
\usage{
excursions_req(n)
}
\arguments{
\item{n}{number of results}
}
\description{
This function returns the number of excursions required to list as impaired for conventional pollutants,
}
\examples{
excursions_req()
}
