% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bact_shell_harvest.R
\name{Shell_Harvest}
\alias{Shell_Harvest}
\title{Bacteria Shell Harvest Analysis}
\usage{
Shell_Harvest(df, criteria = "bact_crit_percent")
}
\arguments{
\item{df}{dataframe with Ecoli data}

\item{criteria}{POSIXCT column name containing criteria values}
}
\value{
a dataframe with relevant Ecoli criteria and excursion variables added
}
\description{
Assesses Ecoli data against the standard
}
\examples{
function(df = your_ecoli_data, criteria = "bact-shell-crit")
}
