% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_assessment.R
\name{temp_assessment}
\alias{temp_assessment}
\title{Temperature Assessment}
\usage{
temp_assessment(df, datetime_column = "sample_datetime",
  spawn_start_column = "spawn_start", spawn_end_column = "spawn_end",
  result_column = "Result_cen", criteria_column = "temp_crit")
}
\arguments{
\item{df}{dataframe with temperature data}

\item{datetime_column}{POSIXCT column name containing sample datetimes}

\item{spawn_start}{mm/dd spawn start date (as characters)}

\item{spawn_end}{mm/dd spawn end date (as characters)}

\item{results}{numeric results column name}

\item{criteria}{name of column containing temperature criteria values}
}
\value{
a dataframe with relevant temperature spawning, criteria, and excursion variables added
}
\description{
Creates three new variables for each sample summarizing the data within the last 3 years at that station; the number of
excursions, the number of samples during the critical period, and the number of samples during the spawning period.
}
