% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP_assessment.R
\name{TP_assessment}
\alias{TP_assessment}
\title{Assess TP data}
\usage{
TP_assessment(TP_data)
}
\value{
dataframe with the following columns added; 'excursion_cen'
}
\description{
Asssesses TP data against the appropriate criteria and produces columns for excursions
}
