% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pH_assessment.R
\name{pH_assessment}
\alias{pH_assessment}
\title{Assess pH data}
\usage{
pH_assessment(pH_data)
}
\arguments{
\item{pH_data}{dataframe with pH data and criteria values}
}
\value{
dataframe with the following columns added; 'pH_violation', 'pH_violation_high', 'pH_violation_low.'
}
\description{
Asssesses pH data against the appropriate criteria and produces columns for excursions, including whether they were high or low
}
\examples{
pH_assessment(pH_data = "your-pH-data")

}
